Refactor data inserter hook management
