+++ workers/loc.api/sync/data.inserter/helpers/index.js
@@ -16,9 +16,9 @@ const {
   getAllowedCollsNames,
   addPropsToResIfExist
 } = require('./utils')
-const convertCurrency = require('./convert-currency')
-const getFlagsFromLedgerDescription = require('./get-flags-from-ledger-description')
-const recalcSubAccountLedgersBalances = require('./recalc-sub-account-ledgers-balances')
+const getFlagsFromLedgerDescription = require(
+  './get-flags-from-ledger-description'
+)
 
 module.exports = {
   searchClosePriceAndSumAmount,
@@ -30,8 +30,6 @@ module.exports = {
   normalizeApiData,
   getAuthFromDb,
   getAllowedCollsNames,
-  convertCurrency,
   addPropsToResIfExist,
-  getFlagsFromLedgerDescription,
-  recalcSubAccountLedgersBalances
+  getFlagsFromLedgerDescription
 }
